FastCDC/requirements.txt
click==8.1.8
click-default-group==1.2.4
codetiming==1.4.0
fastcdc==1.7.0
humanize==4.12.2
kafka-python==2.1.4
py-cpuinfo==9.0.0
